Framerate precision
git-svn-id: svn://svn.code.sf.net/p/mediainfo/code/MediaInfoLib/trunk@1413 45c5ae6f-87cc-4fd0-92ee-e4f023fd80da

+++ Source/MediaInfo/Multiple/File_Riff_Elements.cpp
@@ -2260,7 +2260,7 @@ void File_Riff::SMV0()
         Fill(Stream_Video, 0, Video_MuxingMode, "SMV v1");
         Fill(Stream_Video, 0, Video_Width, Width);
         Fill(Stream_Video, 0, Video_Height, Height);
-        Fill(Stream_Video, 0, Video_FrameRate, FrameRate);
+        Fill(Stream_Video, 0, Video_FrameRate, (float)FrameRate);
         Fill(Stream_Video, 0, Video_FrameCount, FrameCount);
 
         Finished();
